NearbyParentFragment.java: add cache field and method to speed up Drawable retrieval
Before this commit, the ContextCompat.getDrawable() method was extremely slow at retrieving the red
and green map markers. ContextCompat.getDrawable() may be converting the vector marker images to
bitmaps every method call, leading to slow performance.

This commit adds a HashMap field as a cache to NearbyParentFragment.java which maps Contexts and integer ids
to the Drawable object. A private method is added and used to replace ContextCompat.getDrawable(). This new
method utilizes the new cache to greatly increase speed.