NearbyParentFragmentPresenter.kt: set Sitelinks for Nearby Places to Sitelinks from local data
Before this change, when information about a Place was retrieved from the local Place
repository, the Sitelink object was not set into the new Place object. This would
cause the Sitelink icons to be missing from the user interface for some Places.

This change now sets the Sitelinks object to the locally retrieved Place's Sitelinks.
These sitelinks now appear in the user interface if they are stored locally.