NearbyParentFragmentPresenter.kt: refactor indices batching code, simplify channel creation
Code that created batches of indices is now in a separate method (with documentation).
The new method is now called where the replaced code was before this commit. Channel creation
code is simplified by using an unlimited size rather than calculating the number of batches. Some symbols are renamed to increase clarity.