Refactoring NearbyParentFragmentPresenter.loadPlacesDataAsync() (#6586)
* NearbyParentFragmentPresenter.kt: refactor cached place retrieval code

The code which retrieves cached places from the local database is moved
into a separate method. This separate method is then called where the
moved code used to be. Documentation is added to new method.

* NearbyParentFragmentPresenter.kt: refactor indices batching code, simplify channel creation

Code that created batches of indices is now in a separate method (with documentation).
The new method is now called where the replaced code was before this commit. Channel creation
code is simplified by using an unlimited size rather than calculating the number of batches. Some symbols are renamed to increase clarity.

* NearbyParentFragmentPresenter.kt: refactor thread code

The code which each launched thread uses to retrieve information from WikiData
has been moved into a separate method (with documentation). This new method
is then called where the moved code used to be.

* NearbyParentFragmentPresenter.kt: refactor thread launching code

The code which launches the threads that retrieve information from
WikiData has been moved into a separate method (with documentation).
This new method is now called where the moved code used to be.

* NearbyParentFragmentPresenter.kt: refactor WikiData result processing code

The code which processes the WikiData results has been moved into a
separate method (with documentation). This method is called where
the moved code was before this commit.

---------

Co-authored-by: Ritika Pahwa <83745993+RitikaPahwa4444@users.noreply.github.com>